feat(notifications): implement real-time notifications Close #393
