+++ src/notifications/notifications.service.ts
@@ -10,7 +10,7 @@ import {
 } from './entities/notification.entity';
 import { NotificationPreferences } from './entities/notification-preferences.entity';
 import { CreateNotificationDto } from './dto/notification.dto';
-import { NotificationsGateway } from './notifications.gateway';
+import { NotificationsGateway } from '../gateways/notifications/notifications.gateway';
 import { NotificationTemplatesService } from './notification-templates.service';
 import { PreferencesService } from './preferences/preferences.service';
 import { EmailService } from './email/email.service';
